STEP 2: 배송 가격과 가격 할인 정보를 가지는 DeliveryCost Model 추가
